llm-evaluation-app/requirements.txt
flask==3.0.0
boto3==1.34.34
Pillow==10.2.0
python-dotenv==1.0.1